requirements.txt
Flask
pdfplumber
pdf2image
Pillow
opencv-python
numpy
pytesseract
tesseract